댓글 수정 폼

+++ src/main/java/com/jbg/exam/demo/controller/UsrReplyController.java
@@ -1,11 +1,14 @@
 package com.jbg.exam.demo.controller;
 
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import com.jbg.exam.demo.service.ArticleService;
 import com.jbg.exam.demo.service.ReplyService;
 import com.jbg.exam.demo.util.Ut;
+import com.jbg.exam.demo.vo.Article;
 import com.jbg.exam.demo.vo.Reply;
 import com.jbg.exam.demo.vo.ResultData;
 import com.jbg.exam.demo.vo.Rq;
@@ -14,13 +17,16 @@ import com.jbg.exam.demo.vo.Rq;
 public class UsrReplyController {
 
 	private ReplyService replyService;
+	private ArticleService articleService;
 	private Rq rq;
+	
 
-	public UsrReplyController(ReplyService replyService, Rq rq) {
+	public UsrReplyController(ReplyService replyService, ArticleService articleService, Rq rq) {
 		this.replyService = replyService;
+		this.articleService = articleService;
 		this.rq = rq;
 	}
-
+	
 	@RequestMapping("/usr/reply/doWrite")
 	@ResponseBody
 	public String doWrite(String relTypeCode, int relId, String body, String replaceUri) {
@@ -49,20 +55,54 @@ public class UsrReplyController {
 
 		return rq.jsReplace(writeReplyRd.getMsg(), replaceUri);
 	}
-	
-	@RequestMapping("/usr/reply/doDelete")
-	@ResponseBody
-	public String doWrite(int id, String replaceUri) {
+
+	@RequestMapping("/usr/reply/modify")
+	public String modify(int id, String replaceUri, Model model) {
 		if (Ut.empty(id)) {
 			return rq.jsHistoryBack("id(을)를 입력해주세요.");
 		}
 
-		Reply reply = replyService.getReply(id);
+		Reply reply = replyService.getForPrintReply(rq.getLoginedMember(), id);
+
+		if (reply == null) {
+			return rq.historyBackJsOnView(Ut.f("%d번 댓글은 존재하지 않습니다.", id));
+		}
+		if (reply.isExtra__actorCanModify() == false) {
+			return rq.historyBackJsOnView(Ut.f("%d번 댓글을 수정할 권한이 없습니다.", id));
+		}
+		
+		String relDataTitle = null;
+		
+		switch(reply.getRelTypeCode()) {
+		case "article" :
+			Article article = articleService.getArticle(reply.getRelId());
+			relDataTitle = article.getTitle();
+		}
+		
+		model.addAttribute("relDataTitle", relDataTitle);
+		model.addAttribute("reply", reply);
+		
+		
+		return "usr/reply/modify";
+		
+	}
+
+	@RequestMapping("/usr/reply/doDelete")
+	@ResponseBody
+	public String Delete(int id, String replaceUri) {
+		if (Ut.empty(id)) {
+			return rq.jsHistoryBack("id(을)를 입력해주세요.");
+		}
+
+		Reply reply = replyService.getForPrintReply(rq.getLoginedMember(), id);
 
 		if (reply == null) {
 			return rq.jsHistoryBack(Ut.f("%d번 댓글은 존재하지 않습니다.", id));
 		}
 
+		if (reply.isExtra__actorCanDelete() == false) {
+			return rq.jsHistoryBack(Ut.f("%d번 댓글을 삭제할 권한이 없습니다.", id));
+		}
 
 		ResultData deleteReplyRd = replyService.deleteReply(id);
